hasvk: Fix textureGather for R32G32 formats on HSW
Because we need to use ISL_FORMAT_R32G32_FLOAT_LD for use with gather4
for these formats, we need to create separate image views for this.
Currently all image view bindings are duplicated to create a separate
set of image views for gather4 operations when tg4 is used.

Because ISL_FORMAT_R32G32_FLOAT_LD is a float format, we get 1.0 instead
of 1 for SCS_ONE. We fix this by subtracting (bit_cast<uint32_t>(1.0) -
1) from the value when needed. Conditional execution could also be used,
but I couldn't figure that out.

The descriptor limits on image views are now incorrect when
textureGather is used because of creating multiple bindings from a
single descriptor.