vulkan/wsi/wayland: fixes to dma-buf feedback target_device usage
Do not try to use all the tranches from the dma-buf feedback. Ignore
those whose target_device is not interesting for us. This patch was
written with the following rules in mind:

  If we have a non-SCANOUT tranche and its target_device does not match
  the device that we are using, skip it. We are not creating our images
  using such target_device, so it makes no sense to use the parameters
  from this tranche.

  If we have a SCANOUT tranche, things are different. In this case,
  target_device is a KMS capable device and the compositor can take our
  images and use them for direct scanout (if we create our images with
  the parameters advertised in this tranche). So it makes sense to
  consider a SCANOUT tranche even if the target_device does not match
  the device that we are using.

If we can't find a format using the rules above, fallback and consider
all tranches. This is necessary because we still have some issues
regarding compositors reporting the wrong render device in split
display/render SoCs, see
https://gitlab.freedesktop.org/mesa/mesa/-/issues/5591.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>