iris: Extend iris_bo_busy to allow checking for readers/writers
Sometimes it's useful to know whether anyone is reading or writing a
buffer, in addition to simply being in-use (in whatever capacity).

For example, simultaneous reads may be fine.

We add a new 'usage' bitfield allowing the caller to specify whether
to check for busyness due to readers, writers, or both.  We use the
existing PIPE_MAP flags for convenience, even though we aren't exactly
mapping here.  Better than defining yet another set of similar flags.