meson: Don't allow option arrays to be set to auto + explicit values
What currently happens if `-Dgallium-drivers=auto,zink` is that zink is
silently dropped, as `auto` is noticed, and the entire array is replaced
with one that doesn't contain `zink`. The same can happen with
`-Dvulkan-drivers` and `-Dplatforms`.

There are two possible solutions:
1) Error out
2) replace auto with the default list

I've chosen 1 as it's the simplest, and the one that emulates the
behavior we'd really like meson to have for us: that the *default*
values would be set to a sane list for the given OS + architecture. If
that feature ever materializes (ie, the ability to change the default in
the meson_options.txt file based on the machines) we can move to that
seamlessly.

v2:
    - fix copy and paste between drivers and platforms

CC: 22.2 22.3 <mesa-stable@lists.freedesktop.org>
Reviewed-by: Eric Engestrom <eric@igalia.com>
Reviewed-by: Yonggang Luo <luoyonggang@gmail.com>