vbo: track holes in existing bo
And reuse them when possible to avoid allocating tons of memory
when existing objects could be used instead.
This MR reduces VRAM usage of the trace from #6140 by a factor of
approx 2 (compared GALLIUM_HUD=requested-VRAM,stdout output at
frame 20.000 with/without this MR).

BO allocation used to work like this: a single bo (current_bo) was
active and used until there's wasn't enough free space at the end.
Then a new one would be allocated and the logic continues.

The new allocation scheme is similar, but instead of only checking
for available space at the end of the current bo, it checks for free
space in all already allocated BOs.
vbo_save_context keeps a reference to BOs with free space (or holes,
since the "free space must be at the end" limitation has been removed).

To limit global memory usage and performance impact, the number
of holes we keep is limited (to 4 currently).
Also, when vbo_save_context becomes the sole owner of a BO it is
destroyed (= vbo_save_context only keeps BOs that are used by
display lists).

Acked-by: Marek Olšák <marek.olsak@amd.com>