tegra: handle tiling parameters when importing resources
The creation of a resource meant for scanout imposes a linear
tiling so that it can be properly displayed by the hardware,
since no mechanism to inform the kernel of non-trivial tilings
is in place. Remove the linear modifier restriction and add an
IOCTL call to inform the kernel of the correct tiling parameters.

Signed-off-by: Diogo Ivo <diogo.ivo@tecnico.ulisboa.pt>