anv: only prevent the execbuf ioctl when using no_hw in simple_batch
The goal of no_hw should only be to prevent the execbuf ioctl, not the
CPU logic around it. I've even used it to measure CPU overhead for
vm_bind a while ago.

This is not a huge issue here since the simple_batch variation of
these functions is not very important, but having every
anv_gem_execbuffer() guarded by the same exact check will be handy in
the next commits, as this code will change significantly.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>