anv: add a wrapper to anv_gem_execbuffer()
The anv_execbuf layer is completely tied to the execbuffer2 ioctl. If
we want to add support for execbuffer3 (for vm_bind) we need to be
able to support both execbuffer2 and execbuffer3 code paths, and
having a wrapper here will really help us move some of the
ioctl-dependend code to this area, leaving the upper layers a little
more (but not completely) independent of the ioctl version.

Note that we can't wrap all callers of anv_gem_execbuffer() since
query_pass_execbuf code doesn't use the anv_execbuf layer.

Also, let's keep the no_hw checks in the code that immediately that
calls anv_gem_execbuffer().

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>