anv: bring more execbuf2-specific code down to anv_execbuf_submit()
In the future we're going to have execbuf3 in addition to execbuf2, so
in order to minimize the number of checks for 2 vs 3 around the code,
let's continue moving execbuf-version-specific code down to
anv_execbuf_submit() while trying to make struct anv_execbuf a little
more independent of the execbuf version.

When we add execbuf3 support we're going to add an execbuf3 version of
what's currently anv_execbuf_submit().

The 4 new fields added to struct anv_execbuf are what will be needed
for both execbuf2 and execbuf3 ioctls at the anv_execbuf_submit()
layer.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>