anv: only create execbuf->objects during actual command submission
First of all, execbuf->objects is something that's only needed for
execbuffer2 (i.e., won't be needed in execbuffer3), so we don't
actually need to create it before, and creating it during
anv_execbuf_submit() helps us making struct anv_execbuf a little more
independent of the execbuf version.

Second, this array is big and gets constantly reallocated during
creation due to its default size of 64 (which we could/should
increase, by the way). In Dota2, for example, the most common final
size of execbuf->objects is 512 (but I've seen up to 1024 on my
machine), which means we do quite a few reallocations (most of them
involving a copy) of both execbuf->objects and execbuf->bos:
64 -> 128 > 256 -> 512. This change ensures we'll only do the reallocs
with one buffer (anv_execbuf->bos), and then a final malloc of the
correct size for 'objects' when we absolutely need it.

The only thing we use execbuf->objects for during anv_execbuf setup is
to track the usage of EXEC_OBJECT_WRITE. For this we create a new
struct anv_batch_bo_usage that contains a pointer to the buffer and
knows if it's written to or not. We could also go with Iris's approach
and create an array of bits, but that would also result in reallocs
and management of a second array.

One thing I could verify is that with this patch does not make realloc
more likely to return the original pointer (thus saving a copy).

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>