gbm: Fix cursor stride
The stride returned from the CREATE_DUMB ioctl may not
be suitable for the cursor. The CURSOR/CURSOR2 ioctls simply
assume stride==width*4, which is also what a lot of hardware
needs. Populate the cursor bo stride accordingly.

For reference this is the internal ADDFB2 the kernel does
for the CURSOR/CURSOR2 ioctls:
struct drm_mode_fb_cmd2 fbreq = {
	.width = req->width,
	.height = req->height,
	.pixel_format = DRM_FORMAT_ARGB8888,
	.pitches = { req->width * 4 },
	.handles = { req->handle },
};

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>