meson: simplify libxmlconfig
Because we are now sure that `dep_expat.found()` is the same as
`use_xmlconfig` (if we're on an unsupported platform it will be
disabled, and disabling now works correctly), so we don't really need
the variable. Since the dependency is now correctly disabled, we don't
need to conditionally add it to the dependencies, it is either found and
wanted, or not found. This also allows us to pull the preprocessor args
into the library definition

Reviewed-by: Emma Anholt <emma@anholt.net>
Reviewed-by: Eric Engestrom <eric@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/21224>