ac/nir/ngg: Add ability to repack extra output variables.
Use the reusable_repackable option as an upper limit (in dwords)
for repacking and reusing extra output variables.

Currently this means that ac_nir_lower_ngg will try to find outputs
other than the position which are already calculated as part of the
position.

It will then repack these variables after culling (using LDS), and
reuse them to reduce the execution time of the deferred shader part.
This prevents wasting ALU on computing things that have already been
computed as part of the position.

Each repacked-reused dword costs 4 bytes of LDS per vertex and an
additional VGPR used (due to the case when culling is off and
repacking doesn't happen).