ci: fix deleteion of validation-layers directory
We pushd'ed twice, but only popd'ed once, meaning we did not return to
the initial directory. This means we don't actually end up deleting the
directory, and also didn't return to where we started. The latter makes
it harder to write scripts that is executed after this one.

Let's just make the second pushd into a cd instead to fix the problem.

While we're at it, remove the -f flag to rm, which "helfully" hid what
was going on.

There's also no reason to source the script, as we don't export any
environment variables that needs to be picked up here. This alone would
have fixed the issue, but let's do all of these thing in case things
needs to change in the future. In order to do that, we also need to mark
the script as executable.

Fixes: 512f1c160a6 ("ci/zink: Add coverage using the vulkan validation layer on lvp.")