anv/ds: Associate the VkCommandBuffer with the cmd buffer renderstage event.
Like the corresponding tu commit, this means the perfetto UI will have a
non-zero/NULL handle/name in the UI on this renderstage.  Unfortunately,
intel/ds is outside of vulkan so unless we pull in anv and vulkan headers,
we can't just pass in the anv_cmd_buffer.  This also means it would be
much more painful to pass the cmd buffer to the rest of the events, so
they'll still have unset command buffers.  Still, being able to see the
name of the command buffer in at least one of the events should be useful
once that's glued together.