nir/lower_shader_calls: Try to put system values lower
Instead of pulling non-preservable instructions all the way to the top,
we can move them to the lowest common ancestor between them and first call
instruction (except for if statements, where the call inside if-true will
not cause if-else loads to move). This way at least some of them won't
be alive for the whole duration of a function.
Also, as a benefit, now we don't move the instructions that are before
the calls anyway.

And in the worst case, the values will end up on the very top as before.

v2: Added special handling for if statements.

Signed-off-by: Sviatoslav Peleshko <sviatoslav.peleshko@globallogic.com>