spirv2dxil: Lower vars to mediump
This gets a little tricky because we use the precision annotation on
the var to determine how to emit the signature, but:
* This has to match whether the shader is going to treat 16-bit types
  as native or min-precision.
* This has to match across shader stages for pipeline linking.

So, after lowering the vars, we clear the precision annotation (i.e.
lowering to concrete 16-bit instead of min-precision) if the shader
does anything else that'd require native 16-bit types. Then when
linking, if one shader has mediump vars while the other doesn't,
we propagate the native 16-bit property to the next shader.

We also avoid lowering VS inputs and FS int outputs.