mesa: automatically detect gallium-xa
The "gallium-xa" feature cannot be enabled under certain
circumstances, but the meson build detection is unaware of this.  As
a result, configuring mesa on powerpc64le fails with:

  meson.build:663:3: ERROR: Feature gallium-xa cannot be enabled: XA state tracker requires at least one of the following gallium drivers: nouveau, freedreno, i915, svga.

This commit explains to meson.build that if the user has not
explicitly enabled gallium-xa, it should be disabled when the above
condition is not met.