clc: run some LLVM optimization passes on the compiled module
I've seen reduction in size of the final SPIR-V of around 30% with this,
which comes with a couple of benefits for us:

1. Reduce amount of space needed to cache generated SPIR-Vs.
2. Translating between IRs is way faster.
3. `clc_parse_spirv` is less expensive.

There is one potential huge reduction in size and that's by setting
`PTO.MergeFunctions = true;` cutting size of some binaries by 50% on top
of the current benefits. Sadly with this we can end up with pointers to
functions, which without supporting `SPV_INTEL_function_pointers` we can't
pass along to the SPIRV-LLVM-Translator.

Signed-off-by: Karol Herbst <git@karolherbst.de>