gallium/pipe: add query_compute_info
Compute caps has a clunky and unflexible interface which makes it
hard to use especially in Rust code. The kind of of information that
compute caps return is the kind users want to query every cap of at
least once. Which means it doesn't make much sense to have it as a
cap and rather it should be returning a struct.

This allows the code to be more flexible, easier to use, and has less
complexity (especially in Rusticl). This will also enable us to
potentially fix some missing queries or add new previously hard to
implement queries.

v2: rebased, add proper docs, optimized types (Karol)