intel/isl: Treat ISL_FORMAT_YCRCB_* as compressed
The Vulkan 1.3.261 spec says:

    "For the purposes of the constraints on copy extents, this format is
    treated as a compressed format with a 2×1 compressed texel block."

so there's already precident for treating them this way.  Also, that's
really what they are.  It's a very simple 2x1 block compression format
where you do a YUV colorspace conversion as part of the decompression.
Doing this gets rid of a bunch of annoying YUV special casing throughout
ISL.  Also, when we go to copy to/from them, it also means that we now
get R32_UINT copies instead of R16_UINT so it should be faster.