ci/rules: make every job manual in fork pipelines
In forks, every job exists, but no job is automatically run.

Should help with the problem of users clicking on that tempting "play"
button GitLab offers and that has been causing a lot of issues with
wasted resources.

Users now have to use the `bin/ci/ci_run_n_monitor.sh` script to run
a bunch of jobs.