pvr: add VK_EXT_physical_device_drm support
This is notably required by some Wayland compositors implementing
Vulkan renderers.
While the extension is not explicitly clear whether the primary node
should be of the render device or the display device, the consensus
seems to be that it should point to the render device, so this is what
this implementation exposes.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>