ir3: Change merged shader ABI for separable shaders
The VS output ABI and GS input ABI were different depending on what
other shaders were linked in order to avoid reserving rel_patch_id in a
VS followed by a GS where it would be unused. However this won't work
when we compile the VS without knowing which shader follows it. The
slight decrease in register pressure doesn't seem worth the complexity
of plumbing something through to only conditionally change the ABI.

While we're here we merge gs_header and tcs_header, because when we
compile the VS we won't know which one follows.