ir3, tu: Support separable VS and TES
The hardware has a way to compile the VS and TES when we don't know
whether there will be a following shader. We need to do this for shader
objects.

This makes us use this mechanism whenever there may be a following
shader, even if we know for certain that there is. The only penalty I'm
aware of is that this means we always have to compile a binning shader
because key.tess now means "may use tessellation" instead of "will
definitely use tessellation." The alternative would be adding something
to the key and plumbing that through everything.