mesa: Don't recreate RB surfaces for "mismatched" sample counts.
We already for check the underlying pipe_resource changing below, so
there's no need to look inside it here to see if we need to recreate the
surface.  If the chosen samples for the RB changes, that's reflected in
rb->rtt_nr_samples.

Fixes surface recreation overhead in MSRTT situations.