etnaviv: drop ts_size member from etna_resource_level
We only use it to determine if a TS area has been allocated for
the resource level. We don't actually need to track this per
resource level, as TS gets allocated for a resource as a whole,
so we can simply look at the resource containing the level to
check for a TS area.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>