intel/elk: Move code to enable spilling to choose_spill_reg()
The first time we spill a register, we may need to discard and rebuild
the interference graph with some registers reserved, as we unfortunately
have to use registers to send messages to spill registers.  We also need
to set spill costs.

It makes sense to do both of these tasks in choose_spill_reg(), rather
than open coding it in the middle of the spill/retry-allocation loop.
We also introduce a new boolean for whether the current interference
graph supports spilling, in order to simplify some logic.

Cc: mesa-stable
Fixes: e99081e76d4 ("intel/fs/ra: Spill without destroying the interference graph")