nir/lower_io: Support indirect vec components and compact arrays
The tricky part here is to tease out the type_size differences between a
scalar (or array of scalar) variable and a component of a vector.  For
things which are trying to work in terms of GL or Vulkan locations, a
scalar takes up a whole vec4 whenever it's by itself or in a struct or
array.  A component of a vector or a single scalar within a compact
array, however, only take the size of a single scalar.  This adds a
"compact" bool to the type_size callback to allow us to tease out this
difference.

Most drivers should not see a functional change from this because we
only ever set compact when we are handling an array deref of a vector or
when we have an indirect access to clip distance or tess levels.  All
the drivers in the tree today already lower all that away.

Most of this commit is just churning all the drivers to add the new bool
and assert that it's false in most cases.