nir: Add a shared lower_io type size callback which returns vec4s
This is what most drivers want most of the time and it's what gallium
expects.  Most drivers have a type_size function which is a tiny wrapper
like this one in ir3:

    int
    ir3_glsl_type_size(const struct glsl_type *type, bool bindless)
    {
        glsl_count_attribute_slots(type, false);
    }

This provides a common helper which pretty much does that.  The only
difference is that the new helper does ignore images/samplers when
bindless is false so it's safe to use for uniforms, attributes, and
varyings.