mesa,anv,hasvk,turnip,lvp: Call nir_lower_array_deref_of_vec for TCS outputs
TCS outputs can have array derefs of vectors.  We don't have them today
because spirv_to_nir emits a read-select-write pattern.  However, this
is racy if the output is being written from multiple invocations.  We'll
change that in the next but first we need to prep all the tess-capable
drivers.  RADV already invokes nir_lower_array_deref_of_vec in
radv_nir_lower_io_to_scalar_early so it's fine.  The other drivers,
however, need to call the NIR pass.

Notably missing from this list is NVK.  I plan to handle this directly
in the NVK compiler as all I/O is scalar there and we can do indirects
on vectors just fine.