radv: make sure to wait for GDS to be idle with NGG streamout
For some weird reasons, it seems needed to emit a PS_PARTIAL_FLUSH to
fix NGG streamout with Zink on NAVI31.

Note that with Mesa 23.3+, NGG streamout is implemented with GDS
registers for better performance and this workaround is no longer
needed.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/10064
Signed-off-by: Samuel Pitoiset <samuel.pitoiset@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/25971>