freedreno/ci: Skip some uselessly duplicated pipeline tests
Test time for the new CI has vastly increased due to the introduction of
a ton of duplicated pipeline tests for shader objects. Because the tests
were duplicated blindly a lot of them aren't really necessary.

I've went through the list and tried to find the largest test groups
that seem to be testing things that are mostly orthogonal to how we
compile the shaders. This is by no mean comprehensive, but just removing
the biggest groups that iterate over all possible formats and
depth/stencil/blend/etc. state should get us most of the benefit.