anv: add helper to diff against retained instructions
As we emit the internal shader state setup, check if it is different
from the previous programmed state. If that's the case avoid the
emission, otherwise emit it and mark the instruction as dirty for
reemission at the next genX(cmd_buffer_flush_gfx_hw_state) call.

Currently the check always returns true and so there is not
optimization triggered here.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>