meson: Add the possibility to disable libelf support
Change meson to be able to disable libelf support.
The default isn't changed, it will use libelf if it's detected
on the system.
Auto-detect features isn't something that we like on FreeBSD as
it result in different package being built depending on the packages
present on the build system.
Note that on FreeBSD (at least) we will fallback on the system libelf
anyway (which is want we want).

Signed-off-by: Emmanuel Vadot <manu@FreeBSD.org>