nak: Stop converting to CSSA before RA
It's not actually necessary for our implementaiton of Braun/Hack. Early
on, I thought it was a requirement for getting accurate liveness
calculations.  However, because we assume no critical edges, any block
with an OpPhiSrcs has a single successor and the live-out of the block
is never bigger than the live-in of its successor.  This means that, as
long as live-in sets are big enough, we'll still get an accurate live
register count.