radv: Use gang submission for copying images unsupported by SDMA.
When the user tries to execute a transfer queue copy operation
that isn't supported by the SDMA engine, the driver must still
implement that copy somehow. To solve this, we use gang submission
and execute these copies on an async compute queue.

To simplify the implementation and avoid refactoring a large
portion of the code base, we create an internal command
buffer object on which we execute the already existing
compute based copy command implementations.